pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyblaze"
version = "1.5.2"
description = "Jogo de plataforma 2D de alta velocidade inspirado no Sonic"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.scripts]
pyblaze = "pyblaze.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src/pyblaze --cov-report=term-missing"

[tool.coverage.report]
fail_under = 30
omit = [
    "src/pyblaze/main.py",
    "src/pyblaze/scenes/*.py",
    "src/pyblaze/utils/*.py",
    "src/pyblaze/systems/hud.py"
]
